test-scripts/requirements.txt
PyDMX==0.1.0
PyDMX-Drivers-FTDI==0.1.0
pylibftdi==0.22.0
